+++ src/server/start_harvester.py
@@ -10,11 +10,38 @@ except ImportError:
 from src.harvester import Harvester
 from src.server.outbound_message import NodeType
 from src.server.server import ChiaServer
+from src.types.peer_info import PeerInfo
 from src.util.config import load_config, load_config_cli
 from src.util.default_root import DEFAULT_ROOT_PATH
 from src.rpc.harvester_rpc_server import start_harvester_rpc_server
 from src.util.logging import initialize_logging
 from src.util.setproctitle import setproctitle
+
+
+def start_harvester_bg_task(server, peer_info, log):
+    """
+    Start a background task that checks connection and reconnects periodically to the farmer.
+    """
+
+    async def connection_check():
+        while True:
+            farmer_retry = True
+
+            for (
+                connection
+            ) in server.global_connections.get_connections():
+                if connection.get_peer_info() == peer_info:
+                    farmer_retry = False
+
+            if farmer_retry:
+                log.info(f"Reconnecting to farmer {farmer_retry}")
+                if not await server.start_client(
+                    peer_info, None, auth=True
+                ):
+                    await asyncio.sleep(1)
+            await asyncio.sleep(1)
+
+    return asyncio.create_task(connection_check())
 
 
 async def async_main():
@@ -60,7 +87,11 @@ async def async_main():
 
     harvester.set_server(server)
     await asyncio.sleep(1)
-    harvester._start_bg_tasks()
+    peer_info = PeerInfo(
+        config["farmer_peer"]["host"], config["farmer_peer"]["port"]
+    )
+    farmer_bg_task = start_harvester_bg_task(server, peer_info, log)
+
     await server.await_closed()
     harvester._shutdown()
     await harvester._await_shutdown()
@@ -70,6 +101,7 @@ async def async_main():
         await rpc_cleanup()
     log.info("Closed RPC server.")
 
+    farmer_bg_task.cancel()
     log.info("Harvester fully closed.")
